pyproject.toml
[project]
name = "kk"
version = "0.1.0"
description = "kk"
authors = [
    {name = "Eslami, Mahmoud",email = "mahmoud.eslami@students.hs-mainz.de"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "rapidfuzz (>=3.13.0,<4.0.0)",
    "pymongo (>=4.13.2,<5.0.0)",
    "python-telegram-bot (>=22.3,<23.0)",
    "pydantic (>=2.0.0,<3.0.0)",
    "langgraph (>=0.0.40,<1.0.0)",
    "langchain (>=0.1.0,<1.0.0)",
    "langchain-google-genai (>=1.0.0,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
